Did a huge rfacto
