+++ .vscode/launch.json
@@ -1,5 +1,5 @@
 // AUTOMATICALLY GENERATED FILE. PLEASE DO NOT MODIFY IT MANUALLY
-
+//
 // PIO Unified Debugger
 //
 // Documentation: https://docs.platformio.org/page/plus/debugging.html
@@ -12,19 +12,21 @@
             "type": "platformio-debug",
             "request": "launch",
             "name": "PIO Debug",
-            "executable": "e:/sourcen/hivemonitor/.pio/build/esp12e/firmware.elf",
+            "executable": "E:/sourcen/hivemonitor/.pio/build/esp12e/firmware.elf",
+            "projectEnvName": "esp12e",
             "toolchainBinDir": "C:/Users/BajonczakSascha/.platformio/packages/toolchain-xtensa/bin",
+            "internalConsoleOptions": "openOnSessionStart",
             "preLaunchTask": {
                 "type": "PlatformIO",
                 "task": "Pre-Debug"
-            },
-            "internalConsoleOptions": "openOnSessionStart"
+            }
         },
         {
             "type": "platformio-debug",
             "request": "launch",
             "name": "PIO Debug (skip Pre-Debug)",
-            "executable": "e:/sourcen/hivemonitor/.pio/build/esp12e/firmware.elf",
+            "executable": "E:/sourcen/hivemonitor/.pio/build/esp12e/firmware.elf",
+            "projectEnvName": "esp12e",
             "toolchainBinDir": "C:/Users/BajonczakSascha/.platformio/packages/toolchain-xtensa/bin",
             "internalConsoleOptions": "openOnSessionStart"
         }
