+++ platformio.ini
@@ -2,9 +2,10 @@
 
 platform = espressif8266
 board = d1_mini
-
+;board = esp12e
 framework = arduino
-build_flags = -D PIO_FRAMEWORK_ARDUINO_LWIP2_LOW_MEMORY
+build_flags = -D PIO_FRAMEWORK_ARDUINO_LWIP2_LOW_MEMORY -D USE_SCALE_DEVICE=0
+
 lib_deps = Homie 
             ArduinoJson@5.13.4
