+++ src/BatteryProcessor.cpp
@@ -1,0 +1,13 @@
+#include <Homie.h>
+#include "BatteryProcessor.h"
+
+float adjust;
+
+BatteryProcessor::BatteryProcessor()
+{
+}
+
+float BatteryProcessor::getVolt()
+{
+    float aVolatageValue= analogRead(A0);
+}
